+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Allow creating invoices for remote recipients.
 - Send `Offer(Agreement)` activity to start payment process.
+- Implement handler for `Offer(Agreement)` activities.
 
 ### Changed